API refactoring

+++ src/main/java/by/vstu/isit/documentprocessor/services/db/Service.java
@@ -4,6 +4,7 @@ import by.vstu.isit.documentprocessor.excepts.DataNotFoundException;
 import lombok.SneakyThrows;
 import by.vstu.isit.documentprocessor.entities.AbstractEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.Serializable;
 import java.util.List;
@@ -15,26 +16,33 @@ public interface Service<
 
     R getRepository();
 
-    @SneakyThrows()
-    default T getById(ID id) {
-        return getRepository().findById(id).orElseThrow(DataNotFoundException::new);
+    @Transactional(readOnly = true)
+    default T getById(ID id) throws DataNotFoundException {
+        return getRepository()
+                .findById(id)
+                .orElseThrow(() -> new DataNotFoundException("Entity not found: " + id));
     }
 
+    @Transactional(readOnly = true)
     default List<T> getAll() {
         return getRepository().findAll();
     }
 
+    @Transactional
     default T save(T entity) {
         return getRepository().save(entity);
     }
 
-    default boolean delete(T entity) {
+    @Transactional
+    default void delete(T entity) {
         getRepository().delete(entity);
-        return true;
     }
 
-    default boolean delete(ID id) {
+    @Transactional
+    default void delete(ID id) throws DataNotFoundException {
+        if (!getRepository().existsById(id)) {
+            throw new DataNotFoundException("Entity not found: " + id);
+        }
         getRepository().deleteById(id);
-        return true;
     }
 }
